+++ e2e/assets/Ping.bin
@@ -1,0 +1,1 @@
+6080604052348015600e575f5ffd5b5060015f81905550610265806100235f395ff3fe608060405234801561000f575f5ffd5b5060043610610034575f3560e01c80633fa4f24514610038578063773acdef14610056575b5f5ffd5b610040610086565b60405161004d91906100f7565b60405180910390f35b610070600480360381019061006b919061013e565b61008b565b60405161007d91906100f7565b60405180910390f35b5f5481565b5f815f819055507ffd8d0c1dc3ab254ec49463a1192bb2423b3b851adedec1aa94dcd362dc063c9d33836040516100c39291906101a8565b60405180910390a16001826100d891906101fc565b9050919050565b5f819050919050565b6100f1816100df565b82525050565b5f60208201905061010a5f8301846100e8565b92915050565b5f5ffd5b61011d816100df565b8114610127575f5ffd5b50565b5f8135905061013881610114565b92915050565b5f6020828403121561015357610152610110565b5b5f6101608482850161012a565b91505092915050565b5f73ffffffffffffffffffffffffffffffffffffffff82169050919050565b5f61019282610169565b9050919050565b6101a281610188565b82525050565b5f6040820190506101bb5f830185610199565b6101c860208301846100e8565b9392505050565b7f4e487b71000000000000000000000000000000000000000000000000000000005f52601160045260245ffd5b5f610206826100df565b9150610211836100df565b9250828201905080821115610229576102286101cf565b5b9291505056fea26469706673582212204a2be87feda14f693fcf1dfd7921f77c73afa66e24eb41c2f6d6f125dd689f1964736f6c63430008220033
